@@ -7,5 +7,5 @@ using System.Threading.Tasks;
 namespace MobileClient.Logic.Configuration;
 public sealed class ServiceConfig
 {
-    public string MarketService { get; set; }
+    public string MarketService { get; set; } = "http://51.250.31.176:5000";
 }
